text_ocr_dashboard/requirements.txt
streamlit
pillow
numpy
sqlite3
pytesseract
matplotlib
